added proper Decimal support
rewrote some of the rounding utilities to use Decimal, and use it
correctly. this should make rounding more accurate than before.

need to add a set of pint wrapped functions from the math library so
that we can use Decimal types with trig functions, etc, since Decimal does
not work with the numpy module (because it doesn't implement all of the
necessary member, like sin()).